Enforcing compatibility with the librato REST Api
I have had a few problems when using metrics-librato, which inherits the LibratoBatch class. It sometimes pushes NaN values to in the MultiSampleGaugeMeasurement, wich will lead to error 400 when beeing POSTed to the librato API. Even if the real error is located in the metrics-librato package, this class should none the less check for invalid input. The same holds true for the count field: The REST API will not accept fractional counts - thats why i changed it to be a Long.